xcodebuild.go moved under the xcodebuild package;
