+++ xcodebuild/xcodebuild.go
@@ -1,4 +1,4 @@
-package xcodeproj
+package xcodebuild
 
 import (
 	"fmt"
@@ -8,8 +8,7 @@ import (
 	"github.com/bitrise-tools/xcode-project/serialized"
 )
 
-// ParseShowBuildSettingsOutput ...
-func ParseShowBuildSettingsOutput(out string) (serialized.Object, error) {
+func parseShowBuildSettingsOutput(out string) (serialized.Object, error) {
 	settings := serialized.Object{}
 
 	lines := strings.Split(out, "\n")
@@ -37,7 +36,8 @@ func ParseShowBuildSettingsOutput(out string) (serialized.Object, error) {
 	return settings, nil
 }
 
-func showProjectBuildSettings(project, target, configuration, sdk string) (serialized.Object, error) {
+// ShowProjectBuildSettings ...
+func ShowProjectBuildSettings(project, target, configuration, sdk string) (serialized.Object, error) {
 	args := []string{"-project", project, "-target", target, "-configuration", configuration}
 	if sdk != "" {
 		args = append(args, "-sdk", sdk)
@@ -50,5 +50,22 @@ func showProjectBuildSettings(project, target, configuration, sdk string) (seria
 		return nil, fmt.Errorf("%s failed: %s", cmd.PrintableCommandArgs(), err)
 	}
 
-	return ParseShowBuildSettingsOutput(out)
+	return parseShowBuildSettingsOutput(out)
+}
+
+// ShowWorkspaceBuildSettings ...
+func ShowWorkspaceBuildSettings(workspace, scheme, configuration, sdk string) (serialized.Object, error) {
+	args := []string{"-workspace", workspace, "-scheme", scheme, "-configuration", configuration}
+	if sdk != "" {
+		args = append(args, "-sdk", sdk)
+	}
+	args = append(args, "-showBuildSettings")
+
+	cmd := command.New("xcodebuild", args...)
+	out, err := cmd.RunAndReturnTrimmedCombinedOutput()
+	if err != nil {
+		return nil, fmt.Errorf("%s failed: %s", cmd.PrintableCommandArgs(), err)
+	}
+
+	return parseShowBuildSettingsOutput(out)
 }
